% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.data.R
\docType{data}
\name{surv.data}
\alias{surv.data}
\title{Simulated survival data}
\format{
A data frame which contains id, survtime, death, w1.
\describe{
  \item{id}{patients identifier}
  \item{survtime}{survival time (the response variable)}
  \item{death}{censoring indicator, 1=observed, 0=censored}
  \item{w1}{a continoues covariate}
}
}
\usage{
surv.data
}
\description{
Simulated survival data were generated in which "survtime" follows a survival time with a censoring indicator called "death" and a continuous covariate, "w1."
}
\seealso{
\code{\link{ZI}}
}
\keyword{datasets}
