% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZISRE.R
\name{ZISRE}
\alias{ZISRE}
\title{Zero-inflation joint modeling}
\usage{
ZISRE(
  FixedY,
  RandomY,
  GroupY,
  FixedZ,
  RandomZ,
  GroupZ,
  formSurv,
  dataLong,
  dataSurv,
  n.chains = n.chains,
  n.iter = n.iter,
  n.burnin = n.burnin,
  n.thin = n.thin,
  family = "Poisson"
)
}
\arguments{
\item{FixedY}{formula for fixed part of longitudinal count model}

\item{RandomY}{formula for random part of longitudinal count model}

\item{GroupY}{formula specifying the cluster variable for Y (e.g. = ~ subject)}

\item{FixedZ}{formula for fixed part of longitudinal probability model}

\item{RandomZ}{formula for random part of longitudinal probability model}

\item{GroupZ}{formula specifying the cluster variable for Z (e.g. = ~ subject)}

\item{formSurv}{formula for survival model}

\item{dataLong}{data set of observed longitudinal variables.}

\item{dataSurv}{data set of observed survival variables.}

\item{n.chains}{the number of parallel chains for the model; default is 1.}

\item{n.iter}{integer specifying the total number of iterations; default is 1000.}

\item{n.burnin}{integer specifying how many of n.iter to discard as burn-in ; default is 5000.}

\item{n.thin}{integer specifying the thinning of the chains; default is 1.}

\item{family}{Family objects provide a convenient way to specify the details of the models. The families include "Poisson", negative binomial (by "NB") and generalized Poisson (by "GP").}
}
\value{
\itemize{
\item mu.vect list of posterior mean for each parameter
\item sd.vect list of standard error for each parameter
\item 2.5\% list of posterior mode for each parameter
\item 25\% list of posterior median for each parameter
\item 50\% list of posterior median for each parameter
\item 75\% list of posterior median for each parameter
\item 97.5\% list of posterior median for each parameter
\item Rhat Gelman and Rubin diagnostic for all parameter
\item betaL1 the regression coefficients of rate of the hurdle model
\item betaL2 the regression coefficients of probability of the hurdle model
\item betaS the regression coefficients of the survival model
\item Sigma the variance of random effects
\item gamma the association parameters
\item p the scale parameter of Weibull model
\item r the over-dispersion parameter of the negative binomial model
\item phiz the dispersion parameter of the generalized Poisson model
}
}
\description{
Fits zero-inflated hurdle joint models under Poisson, negative binomial, and generalized Poisson distributions by considering shared random effects model.
}
\details{
Function using 'JAGS' software to estimate the hurdle joint modeling
}
\examples{
library(survival)
data(long.data)
data(surv.data)

\donttest{
ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long.data, dataSurv = surv.data,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "Poisson"
)


ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long.data, dataSurv = surv.data,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "NB"
)

ZISRE(
  FixedY = Y1 ~ x1 + x2 + obstime, RandomY = ~obstime, GroupY = ~id,
  FixedZ = ~ x1 + x2 + obstime, RandomZ = ~obstime, GroupZ = ~id,
  formSurv = Surv(survtime, death) ~ w1,
  dataLong = long.data, dataSurv = surv.data,
  n.chains = 2,
  n.iter = 200, n.burnin = 100, n.thin = 1, family = "GP"
)
}
}
\author{
Taban Baghfalaki \email{t.baghfalaki@gmail.com}, Mojtaba Ganjali \email{m-ganjali@sbu.ac.ir}
}
